A bit of cleanup
